refactor(78): merge SavedGame.kt into GameModels.kt
Move SavedGame and InProgressGame into GameModels.kt alongside the other
serialisable domain models and delete the now-empty SavedGame.kt.
